requirements.txt
discord.py
mcstatus 
python-dotenv 
google-cloud-compute
google-auth